[chore] 네이밍, 디렉토리 정리
